Suggestion: parallelize repeat masking
I was trying to build `nt` database and discovered that it would take roughly 12 hours to mask all of it in one thread. This example modification uses GNU Parallel to speed up repeat masking in proportion to the number of CPUs. Perhaps this code will need to be changed before being merged (e. g. checking that GNU Parallel is available in PATH, of taking into account user setting of whether to use multiple threads), so this is only a draft and a suggestion.